+++ CHANGES.md
@@ -1,4 +1,4 @@
-* New puzzle types: Angle Loop, Shikaku.
+* New puzzle types: Angle Loop, Shikaku, Slovak Sums.
 * Add solution code markers.
 * Nurikabe no longer has a spaced format; accepts multidigit
   clues via replacements, as does Fillomino